+++ python/databricks/bundles/pipelines/_models/table_specific_config.py
@@ -80,8 +80,6 @@ class TableSpecificConfig:
     ] = None
     """
     :meta private: [EXPERIMENTAL]
-    
-    (Optional) Additional custom parameters for Workday Report
     """
 
     @classmethod
@@ -150,8 +148,6 @@ class TableSpecificConfigDict(TypedDict, total=False):
     ]
     """
     :meta private: [EXPERIMENTAL]
-    
-    (Optional) Additional custom parameters for Workday Report
     """